% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Cleaning Data}
\usage{
cleanData(x)
}
\arguments{
\item{x}{data frame of at least times and date}
}
\value{
none.
}
\description{
cleaning data
}
\details{
<<<<<<< HEAD
coerces time and date data into proper form
}
\examples{
cleanData(twohundred_fly)

=======
Test.
>>>>>>> 45a48a0 (feedback + proposed changes from cp #1)
}
