% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_event_funct.R
\name{find_event}
\alias{find_event}
\title{Find Event Function}
\usage{
find_event(x, event)
}
\arguments{
\item{x}{data frame of times, date, event, and meet}

\item{event}{the event that the user specifies}
}
\value{
data frame with only the rows containing the selected event, ascending by time
}
\description{
Function that can be used to filter to a specified event
}
\details{
finds all the times corresponding to an event, regardless of year, then sort rows based on best times
}
\examples{
find_event(top_three, "200 Back")
}
