% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_sec_funct.R
\name{make_time_sec}
\alias{make_time_sec}
\title{Making time_sec}
\usage{
make_time_sec(x)
}
\arguments{
\item{x}{dataset}
}
\value{
mutated dataset with time_sec variable
}
\description{
function to make variable called time_sec
}
\details{
outputs a mutated dataset with new variable time_sec (times given only in seconds)
}
\examples{
make_time_sec(clean_twohundred_fly)
}
