% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePeriod.R
\name{makePeriod}
\alias{makePeriod}
\title{making period}
\usage{
makePeriod(min, sec)
}
\arguments{
\item{min}{minutes}

\item{sec}{seconds}
}
\value{
p time in period type
}
\description{
function to make a period with the name of a time unit pluralized
}
\details{
<<<<<<< HEAD
return time of period type
=======
Test.
>>>>>>> 45a48a0 (feedback + proposed changes from cp #1)
}
\examples{
makePeriod(2, 46.70)
}
