% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_meet_type_funct.R
\name{plot_meet_type}
\alias{plot_meet_type}
\title{Plotting performance by meet type}
\usage{
plot_meet_type(x)
}
\arguments{
\item{x}{data frame of at least times, date, and meet name}
}
\value{
none
}
\description{
function to plot different types of meets by times
}
\details{
creates new variable called meet type and shows a line graph for each type
}
\examples{
plot_meet_type(twohundred_fly)
plot_meet_type(twohundred_back)
}
