% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProgress.R
\name{plotProgress}
\alias{plotProgress}
\title{Plotting Progress}
\usage{
plotProgress(x)
}
\arguments{
\item{x}{data frame of at least times and date}
}
\value{
none.
}
\description{
function to plot the progress in one event over the years
}
\details{
cleans the data and plots the race tiems over the years
}
\examples{
plotProgress(twohundred_fly)

}
