% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanData.R
\name{cleanData}
\alias{cleanData}
\title{Cleaning Data}
\usage{
cleanData(x)
}
\arguments{
\item{x}{data frame with at least time and date variables}
}
\value{
none.
}
\description{
cleans data by modifying variables for usability
}
\details{
coerces time and date data into proper format
}
\examples{
cleanData(twohundred_fly)
}
