% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_year_funct.R
\name{filter_year}
\alias{filter_year}
\title{Filter Year Function}
\usage{
filter_year(x, year)
}
\arguments{
\item{x}{data frame of times, date, and meet}

\item{year}{the year that the user specifies}
}
\value{
data frame with only rows of the specified year
}
\description{
Function that can be used to filter to a specified year
}
\details{
input a year and filter out the times from one particular year
}
\examples{
filter_year(twohundred_fly, 2025)
}
