% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_meet_funct.R
\name{find_meet}
\alias{find_meet}
\title{Find Meet Function}
\usage{
find_meet(x, meet)
}
\arguments{
\item{x}{data frame of times, date, and meet}

\item{meet}{the meet that the user specifies}
}
\value{
data frame with only the rows containing the selected meet
}
\description{
Function that can be used to filter to a specified meet
}
\details{
finds all the times from the same meet, regardless of year
}
\examples{
find_meet(twohundred_fly, "Smith vs Simmons")
find_meet(twohundred_fly, "Speedo Sectionals - Providence")
}
