% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePeriod.R
\name{makePeriod}
\alias{makePeriod}
\title{making period}
\usage{
makePeriod(min, sec)
}
\arguments{
\item{min}{minutes}

\item{sec}{seconds}
}
\value{
none.
}
\description{
function to make a period with the name of a time unit pluralized
}
\details{
Test.
}
\examples{
makePeriod(2, 46.70)
}
