% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_avg_funct.R
\name{moving_avg}
\alias{moving_avg}
\title{Creating a moving average}
\usage{
moving_avg(x)
}
\arguments{
\item{x}{data frame of at least times and date}
}
\value{
none
}
\description{
function to plot points with all times and create a moving average
}
\details{
creates new variable called moving_avg and plots it on a graph
}
\examples{
moving_avg(twohundred_fly)
moving_avg(twohundred_back)
}
