% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_progress.R
\name{plot_progress}
\alias{plot_progress}
\title{Plotting Monthly PRs}
\usage{
plot_progress(x)
}
\arguments{
\item{x}{data frame of at least times and date}
}
\value{
none
}
\description{
function to plot best times for each month of each year of competition
}
\details{
cleans the data, converts time to seconds, then plots the times
}
\examples{
plot_progress(twohundred_fly)
}
