% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_pr_funct.R
\name{yearly_pr}
\alias{yearly_pr}
\title{Year PR Function}
\usage{
yearly_pr(x, year)
}
\arguments{
\item{x}{data frame of times, date, and meet}

\item{year}{the year that the user specifies}
}
\value{
one value which is your PR for the specified year
}
\description{
Function that can be used to filter to a specified year and output your PR from that year
}
\details{
finds the PR of a given year
}
\examples{
yearly_pr(twohundred_fly, 2026)
}
